Readme creado, terminado controller, y corrigiendo errores para el postman

+++ apiRestGestionPeliculas/src/main/java/com/salesianos/edu/apirestgestionpeliculas/service/DirectorService.java
@@ -3,9 +3,11 @@ package com.salesianos.edu.apirestgestionpeliculas.service;
 
 import com.salesianos.edu.apirestgestionpeliculas.dto.DirectorRequestDTO;
 import com.salesianos.edu.apirestgestionpeliculas.dto.DirectorResponseDTO;
+import com.salesianos.edu.apirestgestionpeliculas.error.DirectorConPeliculaException;
 import com.salesianos.edu.apirestgestionpeliculas.error.DirectorNotFoundException;
 import com.salesianos.edu.apirestgestionpeliculas.model.Director;
 import com.salesianos.edu.apirestgestionpeliculas.repository.DirectorRepository;
+import com.salesianos.edu.apirestgestionpeliculas.repository.PeliculaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -18,21 +20,20 @@ import java.util.stream.Collectors;
 public class DirectorService {
 
     private final DirectorRepository directorRepository;
+    private final PeliculaRepository peliculaRepository;
 
     public DirectorResponseDTO save(DirectorRequestDTO cmd) {
         return DirectorResponseDTO.fromEntity(directorRepository.save(cmd.toEntity()));
     }
 
-    public List<DirectorResponseDTO> findAll() {
+    public List<Director> findAll() {
         List<Director> result = directorRepository.findAll();
 
         if(result.isEmpty()) {
             throw new DirectorNotFoundException();
         }
 
-        return result.stream()
-                .map(DirectorResponseDTO::fromEntity)
-                .collect(Collectors.toList());
+        return result;
     }
 
     public DirectorResponseDTO findById(Long id) {
@@ -55,6 +56,13 @@ public class DirectorService {
     public void delete(Long id) {
         Director director = directorRepository.findById(id)
                 .orElseThrow(() -> new DirectorNotFoundException(id));
+
+        long peliculasCount = peliculaRepository.countByDirectorId(id);
+
+        if (peliculasCount > 0) {
+            throw new DirectorConPeliculaException(id);
+        }
+
         directorRepository.delete(director);
     }
